refactor(sponsor-reports): fold statusTone into StatusPill
Moves TONE_BY_STATUS + statusTone() from the standalone statusTone.js
into StatusPill.js as a named export. Updates StatusRollupChips.js to
import from StatusPill. Deletes statusTone.js and its test file; all
branch-coverage assertions migrated into StatusPill.test.js.
