+++ src/components/sponsors/reports/StatusPill.js
@@ -1,6 +1,20 @@
 import React from "react";
 import { Chip } from "@mui/material";
-import { statusTone } from "./statusTone";
+
+// Single source of truth: status token -> MUI Chip color. Case-insensitive.
+const TONE_BY_STATUS = {
+  completed: "success",
+  paid: "success",
+  confirmed: "success",
+  pending: "warning",
+  in_progress: "info",
+  not_applicable: "default",
+  canceled: "default",
+  cancelled: "default"
+};
+
+export const statusTone = (status) =>
+  TONE_BY_STATUS[String(status || "").toLowerCase()] || "default";
 
 // A status token rendered as a colored, filled chip. `label` overrides the
 // displayed text (e.g. a T.translate'd label); the color always derives from
@@ -14,5 +28,4 @@ const StatusPill = ({ status, label, size = "small" }) => (
   />
 );
 
-export { statusTone };
 export default StatusPill;
