+++ src/components/sponsors/reports/StatusRollupChips.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Chip, Stack } from "@mui/material";
 import T from "i18n-react/dist/i18n-react";
-import { statusTone } from "./statusTone";
+import { statusTone } from "./StatusPill";
 
 // The backend status_rollup always carries all four lowercase keys; render them
 // in a fixed order so cards line up. A missing rollup degrades to all-zero.
